Added Auto detect + minor changes
Added Auto detect buttons for Steam and GoG.
Code uses windows liberty's to read the registry for the locations.
It has some basic error checking, needs improvement. 
Disable resizing.
Moved the code to move the patch to a function that takes in a string and uses that as a path.